+++ lib/easy_json_matcher.rb
@@ -1,5 +1,5 @@
-require 'json_matcher/node'
-require 'json_matcher/schema_generator'
+require 'easy_json_matcher/node'
+require 'easy_json_matcher/schema_generator'
 module JSONAPIMatcher
 
     SCHEMAS = {}
