+++ lib/easy_json_matcher/boolean_validator.rb
@@ -1,4 +1,4 @@
-require 'json_matcher/validator'
+require 'easy_json_matcher/validator'
 module JSONAPIMatcher
   class BooleanValidator < Validator
     def _validate
