+++ lib/easy_json_matcher/content_wrapper.rb
