+++ lib/easy_json_matcher/node.rb
@@ -1,7 +1,7 @@
-require 'json_matcher'
-require 'json_matcher/validator_factory'
+require 'easy_json_matcher'
+require 'easy_json_matcher/validator_factory'
 require 'json'
-require 'json_matcher/content_wrapper'
+require 'easy_json_matcher/content_wrapper'
 module JSONAPIMatcher
   class Node < Validator
     extend Forwardable
