+++ lib/easy_json_matcher/number_validator.rb
@@ -1,5 +1,5 @@
 # Asserts that a value is a double-precision floating point number in javascript format
-require 'json_matcher/validator'
+require 'easy_json_matcher/validator'
 module JSONAPIMatcher
   class NumberValidator < Validator
     def _validate
