+++ lib/easy_json_matcher/validation_error.rb
