+++ lib/easy_json_matcher/value_validator.rb
@@ -1,4 +1,4 @@
-require 'json_matcher/validator'
+require 'easy_json_matcher/validator'
 module JSONAPIMatcher
   class ValueValidator < Validator
     def _validate
