+++ lib/easy_json_matcher/version.rb
