+++ lib/tasks/jsonapi_matcher_tasks.rake
@@ -1,4 +1,4 @@
 # desc "Explaining what the task does"
-# task :json_matcher do
+# task :easy_json_matcher do
 #   # Task goes here
 # end
